Reply(댓글) Service, ServiceImpl 추가

+++ ex02/src/main/java/org/zerock/mapper/ReplyMapper.java
@@ -10,9 +10,9 @@ public interface ReplyMapper {
 	
 	public int insert(ReplyVO vo);
 	
-	public ReplyVO read(int rno);
+	public ReplyVO read(Long rno);
 	
-	public int delete(int rno);
+	public int delete(Long rno);
 	
 	public int update(ReplyVO reply);
